chore: extract shuffleArray, correct logo treatment
